files: move auth.jsx to auth directory. update import paths for auth.jsx.
